ci: enable ccache for MSVC/Meson, fix warnings, and fix doc workflow trigger (#1359)
* ci: enable ccache for MSVC and Meson, fix MSVC conversion warnings

- Set CMAKE_MSVC_DEBUG_INFORMATION_FORMAT=Embedded (/Z7) when ccache is enabled on MSVC so compilation outputs are cacheable.
- Enable -Db_ccache=true in Meson configure step.
- Fix MSVC truncation/conversion warnings in examples and gridbox_test.

* ci(doc): fix workflow_dispatch placement under on:

* ci: remove unused ccache step and invalid b_ccache flag from test_meson

* ci: fix MSVC debug format for ccache by setting CMP0141 and replacing /Zi with /Z7 after project()

+++ examples/component/flexbox_gallery.cpp
@@ -75,12 +75,12 @@ int main() {
   auto checkbox_group_yflex_grow =
       Checkbox("group |= yflex_grow", &group_yflex_grow);
 
-  auto make_box = [&](size_t dimx, size_t dimy, size_t index) {
+  auto make_box = [&](int dimx, int dimy, int index) {
     std::string title = std::to_string(dimx) + "x" + std::to_string(dimy);
     auto element = window(text(title) | hcenter | bold,
                           text(std::to_string(index)) | hcenter | dim) |
                    size(WIDTH, EQUAL, dimx) | size(HEIGHT, EQUAL, dimy) |
-                   bgcolor(Color::HSV(index * 25, 255, 255)) |
+                   bgcolor(Color::HSV(uint8_t(index * 25), 255, 255)) |
                    color(Color::Black);
     if (element_xflex_grow) {
       element = element | xflex_grow;
